+++ src/particles_base.h
@@ -26,14 +26,14 @@ template<typename CoordT> struct PointMassArray {
         \tparam OtherNumT is a coordinate system of the source PointMassArray */
     template<typename OtherCoordT> PointMassArray(const PointMassArray<OtherCoordT> &src) {
         data.reserve(src.size());
-        for(size_t i=0; i<src.size(); i++)
+        for(unsigned int i=0; i<src.size(); i++)
             data.push_back(ElemType(coord::toPosVel<OtherCoordT,CoordT>(src[i].first), src[i].second));
     }
-    inline ElemType& at(size_t index) { return data.at(index); }             // convenience shorthand
-    inline const ElemType& at(size_t index) const { return data.at(index); } // convenience shorthand
-    inline ElemType& operator[](size_t index) { return data[index]; }        // convenience shorthand
-    inline const ElemType& operator[](size_t index) const { return data[index]; } // another shorthand
-    inline size_t size() const { return data.size(); }  // convenience shorthand
+    inline ElemType& at(unsigned int index) { return data.at(index); }             // convenience shorthand
+    inline const ElemType& at(unsigned int index) const { return data.at(index); } // convenience shorthand
+    inline ElemType& operator[](unsigned int index) { return data[index]; }        // convenience shorthand
+    inline const ElemType& operator[](unsigned int index) const { return data[index]; } // another shorthand
+    inline unsigned int size() const { return data.size(); }  // convenience shorthand
     inline void add( const coord::PosVelT<CoordT> &first, const double second) /// convenience function to add an element
     { data.push_back(ElemType(first, second)); }
 };
